refactor: ThemeService 리팩토링 및 테스트 작성
