+++ src/main/java/roomescape/controller/ThemeController.java
@@ -1,6 +1,7 @@
 package roomescape.controller;
 
 import jakarta.validation.Valid;
+import java.time.LocalDate;
 import java.util.List;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
@@ -43,7 +44,7 @@ public class ThemeController {
     @GetMapping("/themes/famous")
     @ResponseStatus(HttpStatus.OK)
     public List<ThemeResponse> findFamous(@Valid @ModelAttribute ThemeFamousFindRequest request) {
-        List<Theme> themes = themeService.findFamous(request);
+        List<Theme> themes = themeService.findFamous(request, LocalDate.now());
         return themes.stream()
                 .map(ThemeResponse::toDto)
                 .toList();
