+++ src/test/java/roomescape/service/ThemeServiceTest.java
@@ -1,23 +1,37 @@
 package roomescape.service;
 
+import static org.mockito.BDDMockito.given;
+import static org.mockito.Mockito.verify;
+
 import common.exception.RoomEscapeException;
 import java.time.LocalDate;
 import java.util.Optional;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.mockito.InjectMocks;
+import org.mockito.Mock;
+import org.mockito.junit.jupiter.MockitoExtension;
 import roomescape.controller.dto.ThemeFamousFindRequest;
+import roomescape.repository.ReservationRepository;
 import roomescape.repository.ThemeRepository;
-import roomescape.service.stub.StubThemeRepository;
 
+@ExtendWith(MockitoExtension.class)
 public class ThemeServiceTest {
+
+    @Mock
     private ThemeRepository themeRepository;
+
+    @Mock
+    private ReservationRepository reservationRepository;
+
+    @InjectMocks
     private ThemeService themeService;
 
     @Test
     void 존재하지_않는_테마_조회_시_예외가_발생한다() {
         // given
-        themeRepository = new StubThemeRepository(Optional.empty());
-        themeService = new ThemeService(themeRepository);
+        given(themeRepository.findById(999L)).willReturn(Optional.empty());
 
         // when & then
         Assertions.assertThatThrownBy(() -> themeService.find(999L)).isInstanceOf(RoomEscapeException.class);
@@ -26,8 +40,7 @@ public class ThemeServiceTest {
     @Test
     void 존재하지_않는_테마_삭제_시_예외가_발생한다() {
         // given
-        themeRepository = new StubThemeRepository(false);
-        themeService = new ThemeService(themeRepository);
+        given(themeRepository.existsById(999L)).willReturn(false);
 
         // when & then
         Assertions.assertThatThrownBy(() -> themeService.delete(999L)).isInstanceOf(RoomEscapeException.class);
@@ -36,76 +49,66 @@ public class ThemeServiceTest {
     @Test
     void 유명한_테마_조회_시_모든_매개변수가_존재하면_그대로_전달된다() {
         // given
-        StubThemeRepository themeRepository = new StubThemeRepository();
-        themeService = new ThemeService(themeRepository);
-
-        Long days = 10L;
+        long days = 10L;
         LocalDate date = LocalDate.parse("2026-05-01");
-        Long limit = 20L;
+        long limit = 20L;
 
         // when
-        themeService.findFamous(new ThemeFamousFindRequest(days, date, limit));
+        themeService.findFamous(new ThemeFamousFindRequest(days, date, limit), LocalDate.now());
 
         // then
-        Assertions.assertThat(themeRepository.lastCalledDate).isEqualTo(date);
-        Assertions.assertThat(themeRepository.lastCalledDays).isEqualTo(days);
-        Assertions.assertThat(themeRepository.lastCalledLimit).isEqualTo(limit);
+        verify(themeRepository).findFamous(days, date, limit);
     }
 
     @Test
     void 유명한_테마_조회_시_days가_없으면_기본_값으로_대체된다() {
         // given
-        StubThemeRepository themeRepository = new StubThemeRepository();
-        themeService = new ThemeService(themeRepository);
-
-        Long days = null;
         LocalDate date = LocalDate.parse("2026-05-01");
-        Long limit = 20L;
+        long limit = 20L;
 
         // when
-        themeService.findFamous(new ThemeFamousFindRequest(days, date, limit));
+        themeService.findFamous(new ThemeFamousFindRequest(null, date, limit), LocalDate.now());
 
         // then
-        Assertions.assertThat(themeRepository.lastCalledDate).isEqualTo(date);
-        Assertions.assertThat(themeRepository.lastCalledDays).isEqualTo(7L);
-        Assertions.assertThat(themeRepository.lastCalledLimit).isEqualTo(limit);
+        verify(themeRepository).findFamous(7L, date, limit);
     }
 
     @Test
     void 유명한_테마_조회_시_date가_없으면_오늘로_대체된다() {
         // given
-        StubThemeRepository themeRepository = new StubThemeRepository();
-        themeService = new ThemeService(themeRepository);
-
-        Long days = 7L;
-        LocalDate date = null;
-        Long limit = 10L;
-
+        long days = 7L;
+        long limit = 10L;
+        LocalDate now = LocalDate.now();
         // when
-        themeService.findFamous(new ThemeFamousFindRequest(days, date, limit));
+        themeService.findFamous(new ThemeFamousFindRequest(days, null, limit), now);
 
         // then
-        Assertions.assertThat(themeRepository.lastCalledDate).isEqualTo(LocalDate.now());
-        Assertions.assertThat(themeRepository.lastCalledDays).isEqualTo(days);
-        Assertions.assertThat(themeRepository.lastCalledLimit).isEqualTo(limit);
+        verify(themeRepository).findFamous(days, now, limit);
     }
 
     @Test
     void 유명한_테마_조회_시_limit이_없으면_기본값으로_대체된다() {
         // given
-        StubThemeRepository themeRepository = new StubThemeRepository();
-        themeService = new ThemeService(themeRepository);
-
-        Long days = 7L;
+        long days = 7L;
         LocalDate date = LocalDate.parse("2026-05-01");
-        Long limit = null;
 
         // when
-        themeService.findFamous(new ThemeFamousFindRequest(days, date, limit));
+        themeService.findFamous(new ThemeFamousFindRequest(days, date, null), LocalDate.now());
 
         // then
-        Assertions.assertThat(themeRepository.lastCalledDate).isEqualTo(date);
-        Assertions.assertThat(themeRepository.lastCalledDays).isEqualTo(days);
-        Assertions.assertThat(themeRepository.lastCalledLimit).isEqualTo(10L);
+        verify(themeRepository).findFamous(days, date, 10L);
+    }
+
+    @Test
+    void 삭제시_테마가_존재하지_않으면_예외가_발생한다() {
+        given(themeRepository.existsById(999L)).willReturn(false);
+        Assertions.assertThatThrownBy(() -> themeService.delete(999L)).isInstanceOf(RoomEscapeException.class);
+    }
+
+    @Test
+    void 삭제시_테마를_사용하는_예외가_있으면_예외가_발생한다() {
+        given(themeRepository.existsById(1L)).willReturn(true);
+        given(reservationRepository.existsByThemeId(1L)).willReturn(true);
+        Assertions.assertThatThrownBy(() -> themeService.delete(1L)).isInstanceOf(RoomEscapeException.class);
     }
 }
